+++ src/extension.ts
@@ -18,6 +18,34 @@ export function activate(context: vscode.ExtensionContext) {
   const type = WebviewProvider.viewType;
   const sub = vscode.window.registerWebviewViewProvider(type, provider);
 
+  // Status bar item
+  const statusBarItemStart = vscode.window.createStatusBarItem(
+    vscode.StatusBarAlignment.Left,
+    100
+  );
+  statusBarItemStart.command = "easy-odoo-dev.launch-server";
+  statusBarItemStart.text = `$(server) EOD: Start Server`;
+  statusBarItemStart.tooltip = "Launch Odoo Server";
+  statusBarItemStart.show();
+
+  const statusBarItemDebug = vscode.window.createStatusBarItem(
+    vscode.StatusBarAlignment.Left,
+    100
+  );
+  statusBarItemDebug.command = "easy-odoo-dev.debug-server";
+  statusBarItemDebug.text = `$(bug) EOD: Debug Server`;
+  statusBarItemDebug.tooltip = "Debug Odoo Server";
+  statusBarItemDebug.show();
+
+  const statusBarItem = vscode.window.createStatusBarItem(
+    vscode.StatusBarAlignment.Left,
+    100
+  );
+  statusBarItem.command = "easy-odoo-dev.stop-server";
+  statusBarItem.text = `$(stop) EOD: Stop Server`;
+  statusBarItem.tooltip = "Stop Odoo Server";
+  statusBarItem.show();
+
   // Register commands
   const launchServerCommand = vscode.commands.registerCommand(
     "easy-odoo-dev.launch-server",
@@ -48,4 +76,7 @@ export function activate(context: vscode.ExtensionContext) {
   context.subscriptions.push(stopServerCommand);
   context.subscriptions.push(openConfigFileCommand);
   context.subscriptions.push(dropCurrentDatabaseCommand);
+  context.subscriptions.push(statusBarItem);
+  context.subscriptions.push(statusBarItemStart);
+  context.subscriptions.push(statusBarItemDebug);
 }